Use old code style for emacs 23 compat
The new recommended style is to use the "cl-" prefixed versions, but
they do not exist in emacs 23.  We still want to stay compatible with
that, so use plain "loop" rather than "cl-loop" to allow the tests to
work with both old and new versions.

ALso "pcase" was introduced in emacs 24, so stop using it to remain
compatible with emacs 23.