make rust-mode use lexical binding
Emacs 24 introduces lexical binding, which should be preferred for new
code.  This enables it for rust-mode.  The code continues to work fine
on pre-24 Emacs, and it won't be difficult for this to remain true.  One
concrete advantage of lexical binding is that it lets the byte-compiler
generate better warnings in some cases; here it found a couple of unused
variables.