Perform syntactic angle bracket matching
Add angle brackets to the list of emacs matching characters.  Use
syntactic fontification to suppress this where it is not appropriate (in
less than operators, arrows, etc.)  Remove the non-syntactic version
that was there previously.

For electric-pair-mode, suppress the pairing for < and > characters that
are not angle brackets.

Because syntax is used for indentation, this fixes some problems with
it, particularly when attempting to stretch type parameter lists over
multiple lines.