Fix slowness in angle bracket matching
The problem was caused by the logic that made it refontify the whole
buffer when "font-lock-syntactically-fontified" was set to a position
before the start of the region to be potentially fontified.
Unfortunately that variable is not reliably set when fontifying a large
buffer.  Fortunately, the new logic is much simpler, and font-lock
already takes care of ensuring that everything before font-lock-beg was
syntactically fontified.

The other problem was calling syntax-ppss on positions known not to be
fontified yet.

This fixes both of these issues, and the angle bracket matching now
works on larger buffers without pausing on every keystroke.