Recognize runaway raw strings
Recognize raw strings all the way to the end of the buffer if they are
not closed.  This is not valid rust code, but the highlighting should
show the mistake.

This also eliminates glitchy behavior that can occur in this situation.
Emacs assumes that edits can't change syntax at positions before the
edit, and raw strings without this change violated this.