Use syntax-propertize-function, not font-lock-syntactic-keywords
font-lock-syntactic-keywords have been obsolete since Emacs 24.1, the
earliest version supported by rust-mode.  Instead, modes are
encouraged to use syntax-propertize-function.

syntax-propertize-function provides a generally better experience.
Syntax propertization is not tied to font lock,so angle bracket
matching will still work when font-lock is disabled.  There's no
longer a need to call font-lock-fontify-buffer or font-lock-ensure
(you can see this in the tests).  The resulting code is also shorter.

I removed a few tests:

* font-lock-raw-string-constant was written assuming font-lock based
  syntax propertization, but that's no longer the case.
* font-lock-extend-region-in-string and
  rust-test-revert-hook-preserves-point both called functions that no
  longer exist.

There's a fix for a hidden bug in rust-is-in-expression-context.  This
previously could signal in some situations, but the signal was hidden
by font-lock.  It was visible now when running tests, hence the new
call to condition-case.

I suspect this might fix #192, but I haven't tried it.