Pass 'visible to get-buffer-window
If I have two frames, with the *compilation* buffer in one frame and
use next-error in the other frame, I get this error:

Wrong type argument: window-live-p, nil

This happens because get-buffer-window returns nil.  The fix is to
pass 'visible to get-buffer-window, which will pick any visible frame
from the same terminal.