compilation regexps: Match line before, too.
"New format" (since 2016) Rust error messages look like this:

  error[E0308]: mismatched types
     --> src/bin/e19b.rs:672:47
      |
  672 |                 ExprRef::new(0, Nt(NT{ args : others, ..x_nt })),
      |                                               ^^^^^^ expected struct `ExprRef`, found enum `ExprCore`
      |
      = note: expected type `std::vec::Vec<ExprRef>`
		 found type `std::vec::Vec<ExprCore>`

The regexp matches the line with the `-->'.  But of course next-error
scrolls the window so the start of the regexp is at the top of the
buffer.  So we must match the line before, too.  Rust won't ever start
its messages with a `-->' so we can unconditionally include the
previous line in the match.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>