Fix cargo compilation regex starting match
Previously this regex attempted to match the entire cargo output
including the name of the thread that panicked and the entire panic
message. In addition it was required that there be whitespace before
the start of the word "thread" in that output. I don't know if this
was the previous format, but in current Rust compilers "thread" comes
at the beginning of the first line.

However, there were two problems with this approach. One is that
making that leading whitespace mandatory caused the regex to not match
correctly. The second problem is that multiline regex matching in
emacs is a bit iffy, and even matching the start correctly, if the
panic message contained enough characters the full multiline regex
could not be matched and the same problem would occur.

This commit changes and vastly simplifies the regex to search for the
guaranteed portion of the cargo output, starting with the ending
`', `. This does come with an increased risk of accidentally matching
input that is not actually a panic message, but I think in practice
this will be fairly rare.