Establish more-controlled environment for fill-paragraph tests.
Define a macro `rust-test-with-standard-fill-settings` to bind all the
variables that affect `fill-paragraph`, and use it around all calls to
`fill-paragraph`.

Some tests based on `rust-test-fill-paragraph` were failing for me
because I set `sentence-end-double-space` to nil, so double spaces
after periods in the test input were getting collapsed into single
spaces.