rust-doc-indent-line: new function.
This can be used as the indent-line-function, as an alternative to
rust-mode-indent-line.

In addition to indenting the current line as rust-mode-indent-line
does, it also checks to see whether we are editing a Markdown code
block inside a comment (either a documentation comment or a normal
comment.)  If we are inside a Markdown code block, it extracts that
code block, treats it as Rust code, and indents the current line
accordingly.

Note that this behavior is only triggered if point is inside the
comment body: placing point at the start of the line and pressing tab
will only re-indent the comment line, but placing point at the end of
the line and pressing tab will re-indent the both comment line and its
contents.