Set correct rust-buffer-project for remote buffers
rust-buffer-project function calls `cargo locate-project`, which
returns the local path to the Cargo.toml. This causes `cargo` commands
such as `cargo clippy` executed through the rust-mode to fail (because
it is looking for a local path instead of remote host path) if the
project is on a remote tramp connection. Fix by prepending "root" with
remote prefix if `file-remote-p default-directory` is not nil.